refactor: reorganize conveyor components and update import paths
- Moved various conveyor-related components into a dedicated `features/conveyor` directory for improved organization and maintainability.
- Updated import paths across multiple files to reflect this new structure, enhancing clarity and consistency in the codebase.
