+++ src/app/(app)/(stacked-layout)/my-filters/edit/[filterId]/edit-filter-form.tsx
@@ -30,7 +30,7 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { ConveyorCard } from "@/components/conveyor-card";
+import { ConveyorCard } from "@/components/features/conveyor/conveyor-card";
 import { FilterCategoryCombobox } from "@/components/features/my-filters/components/filter-category-combobox";
 import { FormSkeleton } from "@/components/features/my-filters/components/form-skeleton";
 import { FilterImageCombobox } from "@/components/filter-image-combobox";
