+++ src/app/(app)/(stacked-layout)/my-filters/new-filter/new-filter-form.tsx
@@ -24,7 +24,7 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { ConveyorCard } from "@/components/conveyor-card";
+import { ConveyorCard } from "@/components/features/conveyor/conveyor-card";
 import { FilterCategoryCombobox } from "@/components/features/my-filters/components/filter-category-combobox";
 import { FilterImageCombobox } from "@/components/filter-image-combobox";
