+++ src/components/features/conveyor/category-icons.tsx
