+++ src/components/features/conveyor/filter-settings-tooltip.tsx
@@ -3,7 +3,7 @@ import {
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from "./ui/tooltip";
+} from "@/components/ui/tooltip";
 
 interface FilterSettingsTooltipProps {
   children: React.ReactNode;
