+++ src/components/features/filters/filter-card/filter-card-footer.tsx
@@ -9,7 +9,7 @@ import {
   TooltipContent,
   TooltipTrigger,
 } from "@/components/ui/tooltip";
-import { ExportConveyorFilter } from "@/components/export-conveyor-filter";
+import { ExportConveyorFilter } from "@/components/features/conveyor/export-conveyor-filter";
 import ViewFilter from "@/components/features/filters/components/view-filter";
 
 export function FilterCardFooter({
