+++ src/components/features/my-filters/components/bookmarked-filter-card.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import { getR2ImageUrl } from "@/utils/r2-images";
 
 import { ConveyorFilter } from "@/types/filter";
-import { ExportConveyorFilter } from "@/components/export-conveyor-filter";
+import { ExportConveyorFilter } from "@/components/features/conveyor/export-conveyor-filter";
 import ViewFilter from "@/components/features/filters/components/view-filter";
 import { BookmarkToggle } from "@/components/features/filters/filter-card/bookmark-toggle";
